The object memory discarded an _Atomic qualifier five times, quietly
MSVC compiled the whole tree this time and reached the link.  Four warnings
in that output were worth reading rather than scrolling past, and two of them
were bugs.

C4090, "different '_Atomic' qualifiers", five times across om_mt.c and
image_mt.c.  st_om_table and st_om_refcounts point at _Atomic objects, and C
does not let a qualifier drop in silence on the way to void * -- the
conversion is a constraint violation and a compiler may say so.  gcc never
has, through every -Wall -Wextra -Wpedantic build this tree has ever had.
Nothing was broken: free and realloc want the allocation and not the atomics
in it, and the one memset runs during image load with no worker started.  But
five silent qualifier discards in the object memory of a threaded system is
not a line anyone should learn to scroll past, so they are casts now.  The
convention was already in the file -- memset((void *) worker_epoch, ...) --
and the other five had simply never been asked.

C4244, uint32_t to uint16_t, in -inspect.  A real bug and a small one.
OM_fetch_word_length answers uint32_t; it was held in a uint16_t, which
truncated BEFORE the `if (n > 16)' clamp two lines later.  An object of
exactly 65536 words became zero, so the diagnostic printed no fields at all
for it -- the single size of object it cannot describe, reported as empty.
Under OM=mt an Array that big is ordinary.  Both n and i are uint32_t now.

C4267, size_t to unsigned, in discard_statement_value.  Not a bug: a method's
bytecodes do not approach 4 GB.  Cast, so the narrowing is stated rather than
implied.

C4701, "potentially uninitialized local variable 'receiver' used", twice in
compiler.c.  NOT a bug and deliberately not silenced.  receiver is read only
under `if (sent)', and sent is set only after mark(c, &receiver) has run;
gcc's flow analysis correlates the two and MSVC's does not.  Initialising it
to quiet the warning would hide the correlation from the next reader as well
as from the compiler.  Windows.md records it as expected noise instead, with
C4456 and C4457, which are shadowed declarations and equally not defects.

LNK4031, "no subsystem specified; CONSOLE assumed" -- a correct guess, now
named.  /SUBSYSTEM:CONSOLE goes after /link and after the inputs, which is
the only place cl passes a linker option through instead of answering D9002
for an option it has none of.  Every test link gets it too.

Two things the same build settled that were open questions in Windows.md.
SDL3= finds the import library at lib\x64\SDL3.lib, which is where
libsdl.org's SDL3-devel-3.4.14-VC package puts it -- so SDL3LIB= is for other
layouts and not for that one.  And shell32.lib is not needed: SDL_main.h's
entry point resolves without it.  Both were guesses in this file and are now
answers.

Windows.md stops saying the build has not finished there, because it now
compiles every translation unit and links.  What is still untested is
everything after that: no image has been bootstrapped on Windows, no window
opened, no suite run.  The document now draws the line at the link rather
than at the compiler, and says the Win32 half of the prim.c file shim --
_O_BINARY, _chsize_s, _filelengthi64, OVERLAPPED -- is exercised by nothing
yet and is the first place to look if an image written there reads back
short.

Both memories build warning-clean on Linux; all sixteen unit suites and all
five imported-package suites pass at their recorded scores; -inspect and
-census verified by hand after the type change.

+++ src/compiler/compiler.c
@@ -2800,7 +2800,10 @@ discard_statement_value(st_compiler *c)
      *  the current end means these bytes are a target and must not move.
      */
     if (c->store_end == c->out->length && c->store_end != NO_STORE) {
-        c->out->length = c->store_at;
+        /*  store_at is a size_t and length is unsigned; a method's bytecodes
+         *  do not approach 4 GB, so the cast is the narrowing said out loud
+         *  rather than a bound being asserted.  MSVC C4267.  */
+        c->out->length = (unsigned) c->store_at;
         c->store_end   = NO_STORE;
         switch (c->store_kind) {
         case STORE_TEMPORARY:
